+++ src/Oro/Bundle/MessageQueueBundle/DependencyInjection/Configuration.php
@@ -53,6 +53,36 @@ class Configuration implements ConfigurationInterface
             ->end()
             ->arrayNode('persistent_processors')
                 ->prototype('scalar')->end()
+            ->end()
+            ->arrayNode('time_before_stale')
+                ->example("
+                    time_before_stale:
+                        default: X
+                        jobs:
+                            some_job_type_name: Y
+                    ")
+                ->children()
+                    ->integerNode('default')
+                        ->min(-1)
+                        ->info('Number of seconds of inactivity to qualify job as stale. 
+                            If this attribute is not set or it is set to -1 jobs will never be qualified as stale. 
+                            It means that if a unique Job is not properly removed after finish it will be blocking other
+                            Jobs of that type, until it will be manually interrupted')
+                    ->end()
+                    ->arrayNode('jobs')
+                        ->useAttributeAsKey('job_name')
+                        ->info('Number of seconds of inactivity to qualify jobs of this type as stale.
+                                To disable staling jobs for given job type set this option to -1. 
+                                Key can be whole job name or a part of it from the beginning of string to any "."')
+                        ->example("
+                            jobs:
+                                bundle_name.processor_name.entity_name.user: X
+                                bundle_name.processor_name.entity_name: Y
+                                bundle_name.processor_name: Z
+                            ")
+                        ->prototype('integer')->end()
+                    ->end()
+                ->end()
             ->end();
 
         return $tb;