+++ src/Oro/Bundle/MessageQueueBundle/Migrations/Schema/OroMessageQueueBundleInstaller.php
@@ -23,7 +23,7 @@ class OroMessageQueueBundleInstaller implements Installation, ContainerAwareInte
      */
     public function getMigrationVersion()
     {
-        return 'v1_5';
+        return 'v1_5_1';
     }
 
     /**
@@ -88,6 +88,7 @@ class OroMessageQueueBundleInstaller implements Installation, ContainerAwareInte
         $table->addColumn('created_at', 'datetime');
         $table->addColumn('started_at', 'datetime', ['notnull' => false]);
         $table->addColumn('stopped_at', 'datetime', ['notnull' => false]);
+        $table->addColumn('last_active_at', 'datetime', ['notnull' => false]);
         $table->addIndex(['owner_id'], "owner_id_idx");
         $table->setPrimaryKey(['id']);
         $table->addColumn('data', 'json_array', [