+++ src/Oro/Component/MessageQueue/Job/Job.php
@@ -9,6 +9,12 @@ class Job
     const STATUS_FAILED = 'oro.message_queue_job.status.failed';
     const STATUS_FAILED_REDELIVERED = 'oro.message_queue_job.status.failed_redelivered';
     const STATUS_CANCELLED = 'oro.message_queue_job.status.cancelled';
+    const STATUS_STALE = 'oro.message_queue_job.status.stale';
+
+    /**
+     * @var array
+     */
+    public static $activeStatuses = [self::STATUS_NEW, self::STATUS_RUNNING, self::STATUS_FAILED_REDELIVERED];
 
     /**
      * @var int
@@ -59,6 +65,11 @@ class Job
      * @var \DateTime
      */
     protected $startedAt;
+
+    /**
+     * @var \DateTime
+     */
+    protected $lastActiveAt;
 
     /**
      * @var \DateTime
@@ -269,6 +280,25 @@ class Job
     public function setStartedAt(\DateTime $startedAt)
     {
         $this->startedAt = $startedAt;
+    }
+
+    /**
+     * @return \DateTime
+     */
+    public function getLastActiveAt()
+    {
+        return $this->lastActiveAt;
+    }
+
+    /**
+     * @param \DateTime $lastActiveAt
+     *
+     * @return $this
+     */
+    public function setLastActiveAt(\DateTime $lastActiveAt)
+    {
+        $this->lastActiveAt = $lastActiveAt;
+        return $this;
     }
 
     /**